introduce SysInfo object for cpus and uptime methods
