+++ failgood/common/src/failgood/internal/SysInfo.kt
@@ -1,0 +1,7 @@
+package failgood.internal
+
+expect object SysInfo {
+    internal fun upt(): Long
+    internal fun uptime(totalTests: Int? = null): String
+    fun cpus(): Int
+}
