+++ failgood/js/src/failgood/internal/SysInfo.js.kt
@@ -1,0 +1,15 @@
+package failgood.internal
+
+actual object SysInfo {
+    internal actual fun upt(): Long {
+        TODO("Not yet implemented")
+    }
+
+    internal actual fun uptime(totalTests: Int?): String {
+        TODO("Not yet implemented")
+    }
+
+    actual fun cpus(): Int {
+        TODO("Not yet implemented")
+    }
+}
