+++ failgood/jvm/test/failgood/experiments/CoroutinesTest.kt
@@ -1,6 +1,6 @@
 package failgood.experiments
 
-import failgood.uptime
+import failgood.internal.SysInfo.uptime
 import kotlinx.coroutines.asCoroutineDispatcher
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
